projects/etl-pipeline-demo/requirements.txt
pandas
pyyaml
sqlalchemy
apache-airflow
python-dotenv
